# Simplify `Rows::drop()` method