--- a/.github/workflows/test-benchmark.yml
+++ b/.github/workflows/test-benchmark.yml
@@ -78,7 +78,7 @@ jobs:
             echo '<details><summary>Extractors</summary>'
             echo '  '
             echo '```shell'
-            composer test:benchmark -- --ref=1.x --progress=none --group=extractor
+            composer test:benchmark:extractor -- --ref=1.x --progress=none
             echo '```'
             echo '  '
             echo '</details>'
@@ -86,7 +86,7 @@ jobs:
             echo '<details><summary>Transformers</summary>'
             echo '  '
             echo '```shell'
-            composer test:benchmark -- --ref=1.x --progress=none --group=transformer
+            composer test:benchmark:transformer -- --ref=1.x --progress=none
             echo '```'
             echo '  '
             echo '</details>'
@@ -94,7 +94,7 @@ jobs:
             echo '<details><summary>Loaders</summary>'
             echo '  '
             echo '```shell'
-            composer test:benchmark -- --ref=1.x --progress=none --group=loader
+            composer test:benchmark:loader -- --ref=1.x --progress=none
             echo '```'
             echo '  '
             echo '</details>'
@@ -102,7 +102,7 @@ jobs:
             echo '<details><summary>Building Blocks</summary>'
             echo '  '
             echo '```shell'
-            composer test:benchmark -- --ref=1.x --progress=none --group=building_blocks
+            composer test:benchmark:building_blocks -- --ref=1.x --progress=none
             echo '```'
             echo '  '
             echo '</details>'