--- a/src/core/etl/tests/Flow/ETL/Tests/Benchmark/RowsBench.php
+++ b/src/core/etl/tests/Flow/ETL/Tests/Benchmark/RowsBench.php
@@ -31,4 +31,12 @@ final class RowsBench
 
         }
     }
+
+    #[Revs(5)]
+    public function bench_drop_100() : void
+    {
+        foreach ($this->rows->drop(100) as $chunk) {
+
+        }
+    }
 }