--- a/src/core/etl/tests/Flow/ETL/Tests/Benchmark/RowsBench.php
+++ b/src/core/etl/tests/Flow/ETL/Tests/Benchmark/RowsBench.php
@@ -25,7 +25,7 @@ final class RowsBench
     }
 
     #[Revs(5)]
-    public function bench_chunk_10() : void
+    public function bench_chunk_10_on_10k() : void
     {
         foreach ($this->rows->chunks(10) as $chunk) {
 
@@ -33,7 +33,7 @@ final class RowsBench
     }
 
     #[Revs(5)]
-    public function bench_drop_100() : void
+    public function bench_drop_100_on_10k() : void
     {
         foreach ($this->rows->drop(100) as $chunk) {