<!-- Bellow section will be used to automatically generate changelog, please do not modify HTML code structure -->
<h2>Change Log</h2>
<div id="change-log">
  <h4>Added</h4>
  <ul id="added">
    <!-- <li>Feature making everything better</li> -->
  </ul> 
  <h4>Fixed</h4>  
  <ul id="fixed">
    <!-- <li>Behavior that was incorrect</li> -->
  </ul>
  <h4>Changed</h4>
  <ul id="changed">
    <li>Simplify `Rows::drop()` method</li>
    <li>Rework benchmark scripts in composer.json</li>
  </ul>  
  <h4>Removed</h4>
  <ul id="removed">
    <!-- <li>Something</li> -->
  </ul>
  <h4>Deprecated</h4>
  <ul id="deprecated">
    <!-- <li>Something is from now deprecated</li> -->
  </ul>  
  <h4>Security</h4> 
  <ul id="security">
    <!-- <li>Something that was a security issue, is not an issue anymore</li> -->
  </ul>     
</div>
<hr/>

<h2>Description</h2>

Old:
```console
+-------------------------+----------------+------+-----+----------+-----------+
| benchmark               | subject        | revs | its | mem_peak | mode      |
+-------------------------+----------------+------+-----+----------+-----------+
| RowsBench               | bench_chunk_10 | 5    | 5   | 57.886mb | 2.538ms   |
| RowsBench               | bench_drop_100 | 5    | 5   | 57.886mb | 12.008ms  |
| NativeEntryFactoryBench | bench_10k_rows | 5    | 5   | 84.180mb | 130.308ms |
| NativeEntryFactoryBench | bench_1k_rows  | 5    | 5   | 45.708mb | 13.086ms  |
| NativeEntryFactoryBench | bench_5k_rows  | 5    | 5   | 62.981mb | 65.004ms  |
+-------------------------+----------------+------+-----+----------+-----------+
```
New:
```console
+-------------------------+----------------+------+-----+----------+-----------+
| benchmark               | subject        | revs | its | mem_peak | mode      |
+-------------------------+----------------+------+-----+----------+-----------+
| RowsBench               | bench_chunk_10 | 5    | 5   | 57.886mb | 2.558ms   |
| RowsBench               | bench_drop_100 | 5    | 5   | 57.886mb | 4.639ms   |
| NativeEntryFactoryBench | bench_10k_rows | 5    | 5   | 84.180mb | 130.317ms |
| NativeEntryFactoryBench | bench_1k_rows  | 5    | 5   | 45.699mb | 12.981ms  |
| NativeEntryFactoryBench | bench_5k_rows  | 5    | 5   | 62.991mb | 63.155ms  |
+-------------------------+----------------+------+-----+----------+-----------+
```